requirements.txt
requests==2.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings>=2.2
sqlalchemy==2.0.25
alembic==1.13.1
dbfread==2.0.7


